backend/OctoRadarServer/requirements.txt
toml==0.10.2
cerberus==1.3.5
pprintpp==0.4.0
setproctitle==1.3.3
colorlog==6.7.0
pymongo
flask
python-dotenv